reorganized folder structure and deleted unnecessary files. Not sure what to do with VoteButton.js at ./Components, doesn't look like it's being used.
