+++ src/Components/Extension/SinglePullRequestView/VoteText.js
