+++ src/Components/Extension/SinglePullRequestView/VoteTotal.js
